Two Sample T-test Comparisons

This function performs pairwise comparisons between two groups for each combination of a categorical predictor (with exactly two levels) and a continuous outcome variable. It first converts any character variables in x.df to factors and applies a log2 transformation to the continuous variables if specified. Depending on the value of scale, the function conducts either a two-sample t-test or a Mann-Whitney U test and prints the resulting p-values. An error is thrown if a categorical variable does not have exactly two levels.

Usage

cyt.ttests(x.df, scale = NULL)

Arguments

x.df

A matrix or data frame containing continuous variables and categorical variables.

scale

A character value specifying a transformation for continuous variables. Options are NULL (default) and "log2". When scale = "log2", a log2 transformation is applied and a two-sample t-test is used; when scale is NULL, a Mann-Whitney U test is performed.

Value

A list of p-values from the statistical tests for each combination of continuous outcome and categorical predictor is returned.

Examples